fix(tests): move real Blob geometry checks out of patched generation scope
